Improve error handling, tests and typing of PeriodicSaveCallback (#143)

+++ poutyne/framework/callbacks/periodic.py
@@ -110,7 +110,7 @@ class PeriodicSaveCallback(ABC, Callback):
         atomic_write: bool = True,
         open_mode: str = 'wb',
         read_mode: str = 'rb',
-    ):
+    ) -> None:
         super().__init__()
         self.filename = filename
         self.monitor = monitor
@@ -142,7 +142,7 @@ class PeriodicSaveCallback(ABC, Callback):
         self.period = period
 
     @abstractmethod
-    def save_file(self, fd: IO, epoch_number: int, logs: Dict):
+    def save_file(self, fd: IO, epoch_number: int, logs: Dict) -> None:
         """
         Abstract method that is called every time a save needs to be done.
 
@@ -153,7 +153,7 @@ class PeriodicSaveCallback(ABC, Callback):
         """
         pass
 
-    def _save_file(self, filename: str, epoch_number: int, logs: Dict):
+    def _save_file(self, filename: str, epoch_number: int, logs: Dict) -> None:
         atomic_lambda_save(
             filename,
             self.save_file,
@@ -163,35 +163,38 @@ class PeriodicSaveCallback(ABC, Callback):
             atomic=self.atomic_write,
         )
 
-    def on_epoch_end(self, epoch_number: int, logs: Dict):
+    def on_epoch_end(self, epoch_number: int, logs: Dict) -> None:
         filename = self.filename.format_map(logs)
 
         if self.save_best_only:
-            if self.monitor_op(logs[self.monitor], self.current_best):
-                old_best = self.current_best
-                self.current_best = logs[self.monitor]
-                old_best_filename = self.best_filename
-                self.best_filename = filename
+            if self.monitor in logs:
+                if self.monitor_op(logs[self.monitor], self.current_best):
+                    old_best = self.current_best
+                    self.current_best = logs[self.monitor]
+                    old_best_filename = self.best_filename
+                    self.best_filename = filename
 
-                if self.verbose:
-                    print(
-                        f'Epoch {epoch_number:d}: {self.monitor} improved from {old_best:0.5f} '
-                        f'to {self.current_best:0.5f}, saving file to {self.best_filename}'
-                    )
-                self._save_file(self.best_filename, epoch_number, logs)
-                if (
-                    self.keep_only_last_best
-                    and self.best_filename != old_best_filename
-                    and old_best_filename is not None
-                ):
-                    os.remove(old_best_filename)
+                    if self.verbose:
+                        print(
+                            f'Epoch {epoch_number:d}: {self.monitor} improved from {old_best:0.5f} '
+                            f'to {self.current_best:0.5f}, saving file to {self.best_filename}'
+                        )
+                    self._save_file(self.best_filename, epoch_number, logs)
+                    if (
+                        self.keep_only_last_best
+                        and self.best_filename != old_best_filename
+                        and old_best_filename is not None
+                    ):
+                        os.remove(old_best_filename)
+            else:
+                raise KeyError(f"The monitored metric name {self.monitor} is not found in computed metrics.")
         elif epoch_number % self.period == 0:
             if self.verbose:
                 print(f'Epoch {epoch_number:d}: saving file to {filename}')
             self._save_file(filename, epoch_number, logs)
 
     @abstractmethod
-    def restore(self, fd: IO):
+    def restore(self, fd: IO) -> None:
         """
         Abstract method that is called when a save needs to be restored. This happens at the end of the training when
         ``restore_best`` is true.
@@ -201,7 +204,7 @@ class PeriodicSaveCallback(ABC, Callback):
         """
         pass
 
-    def on_train_end(self, logs: Dict):
+    def on_train_end(self, logs: Dict) -> None:
         if self.restore_best:
             if self.best_filename is not None:
                 if self.verbose:
@@ -229,14 +232,14 @@ class PeriodicSaveLambda(PeriodicSaveCallback):
         :class:`~poutyne.PeriodicSaveCallback`
     """
 
-    def __init__(self, func: Callable, *args, restore: Optional[Callable] = None, **kwargs):
+    def __init__(self, func: Callable, *args, restore: Optional[Callable] = None, **kwargs) -> None:
         super().__init__(*args, **kwargs)
         self.func = func
         self._restore = restore
 
-    def save_file(self, fd: IO, epoch_number: int, logs: Dict):
+    def save_file(self, fd: IO, epoch_number: int, logs: Dict) -> None:
         self.func(fd, epoch_number, logs)
 
-    def restore(self, fd: IO):
+    def restore(self, fd: IO) -> None:
         if self._restore is not None:
             self._restore(fd)
